social_media_api/requirements.txt
asgiref==3.8.1
Django==5.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-taggit==6.0.0
django_csp==3.8
djangorestframework==3.15.2
mysqlclient==2.2.4
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
